feat(floating-button-overhaul): buttons split with minimal changes
